"Refactored project structure and added frontend
Reworked the structure of the project to segregate frontend and backend code. This provides better organization and separation of concerns. The backend was divided into modules and handlers for better modularity.

Added a frontend with a HTML interface to interact with the application. User can now input a zip code, and the application will fetch and display information about it. The frontend was built with the help of bootstrap and jquery libraries, which were added to the project.

Modified handlers and tests to adapt to the new project structure."