Add jsLibraryMappings and .gitignore files
Added jsLibraryMappings.xml file to map the JavaScript libraries used in the project. This will allow IDEs to better understand the project structure and provide relevant auto-completions. Also, included a .gitignore file to prevent unneeded files from being tracked by Git. This helps to keep our repository clean from unnecessary build and temporary files which are not needed for the project to function.